refactor: usePageVisibility 훅을 WebSocketContext가 아닌 useSyncExternalStore를 사용해 전역 상태로 수정
- 기존 usePageVisibility 훅을 useSyncExternalStore를 사용한 새로운 구현으로 대체하여 웹소켓 관련 코드에서 가독성 향상
- WebSocketContext 및 WebSocketProvider에서 isVisible 속성을 제거하고, useWebSocketReconnection 및 useWebSocketSubscription에서 직접 호출하도록 수정
