+++ frontend/src/apis/websocket/contexts/WebSocketProvider.tsx
@@ -1,20 +1,16 @@
 import { PropsWithChildren } from 'react';
-import { usePageVisibility } from '../hooks/usePageVisibility';
 import { useWebSocketConnection } from '../hooks/useWebSocketConnection';
 import { useWebSocketMessaging } from '../hooks/useWebSocketMessaging';
 import { useWebSocketReconnection } from '../hooks/useWebSocketReconnection';
 import { WebSocketContext, WebSocketContextType } from './WebSocketContext';
 
 export const WebSocketProvider = ({ children }: PropsWithChildren) => {
-  const { isVisible } = usePageVisibility();
-
   const { client, isConnected, startSocket, stopSocket } = useWebSocketConnection();
 
   const { subscribe, send } = useWebSocketMessaging({ client, isConnected });
 
   useWebSocketReconnection({
     isConnected,
-    isVisible,
     startSocket,
     stopSocket,
   });
@@ -25,7 +21,6 @@ export const WebSocketProvider = ({ children }: PropsWithChildren) => {
     subscribe,
     send,
     isConnected,
-    isVisible,
     client,
   };
